// frontend/src/types.ts
export type Icon =
  | 'mobile'
  | 'credit'
  | 'other_payment'
  | 'hobby'
  | 'subscription'
  | 'transport'
  | 'restaurant'
  | 'utility'
  | 'shopping'
  | 'debt';

export type Category =
  | 'mobile'
  | 'credit'
  | 'other_payment'
  | 'hobby'
  | 'subscription'
  | 'transport'
  | 'restaurant'
  | 'utility'
  | 'shopping'
  | 'debt';